requirements.txt
aiogram>=3.7.0,<4.0
aiohttp>=3.9.0,<4.0
aiosqlite>=0.20.0
apscheduler>=3.10.0,<4.0
loguru>=0.7.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
redis>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
